Rename classify_shading_vs_degradation.R to classification.R
This code is used to classify whether a panel is performing as expected or is showing reduced performance and then classifies it into one of two failure modes (shading and degradation). Shading is defined by reduced group mean with the panels around it but low variability. Degradation is defined as high variability with one panel indicating that the panel itself is degrading.